+++ src/main/java/com/overclock/overclock/dao/OverclockDAO.java
@@ -2,8 +2,10 @@ package com.overclock.overclock.dao;
 
 import com.overclock.overclock.model.Overclock;
 
+import java.math.BigInteger;
+
 public interface OverclockDAO {
-    Overclock getOverclockById(int id);
+    Overclock getOverclockById(BigInteger id);
     boolean save(Overclock overclock);
-    boolean update(int id, Overclock newOverclock);
+    boolean update(BigInteger id, Overclock newOverclock);
 }
